added data tracking, with basic check that the value from cache is correct
The cache "line" defined is not clear. The readSize/writeSize
parameters are measure of the number of bytes per memory operation. The
"line" size is the number of bytes in a cache line. If an access bleeds
across cache lines, it calls Access instead of AccessSingleLine. This
occurs when the memory operation is larger than the size of a cache line
(64 bytes). Why is the Single/non-Single access limit set at 4? a 4 byte
cache line makes no sense??